third_party/blink/web_tests/external/wpt/pointerevents/pointerevent_after_target_removed_from_slot_mouse-expected.txt

This is a testharness.js-based test.
[FAIL] mouse events with remove-slot at pointerdown
  assert_equals: events received after removal expected "pointerout@filler,pointerleave@filler,pointerover@parent,pointerover@host,pointerup@parent,pointerup@host,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerout@host,pointerleave@parent,pointerleave@host" but got "pointerout@filler,pointerleave@filler,pointerover@parent,pointerover@host,pointerenter@host,pointerenter@parent,pointerup@parent,pointerup@host,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerout@host,pointerleave@parent,pointerleave@host"
[FAIL] mouse events with remove-filler at pointerdown
  assert_equals: events received after removal expected "pointerover@slot,pointerup@slot,pointerdown@slot,pointerup@slot,pointerout@slot,pointerleave@slot,pointerleave@parent,pointerleave@host" but got "pointerover@parent,pointerenter@parent,pointerup@parent,pointerup@host,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerout@host,pointerleave@parent,pointerleave@host"
[FAIL] mouse events with change-slotname at pointerdown
  assert_equals: events received after removal expected "pointerout@filler,pointerleave@filler,pointerover@parent,pointerover@host,pointerup@parent,pointerup@host,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerout@host,pointerleave@parent,pointerleave@host" but got "pointerout@filler,pointerleave@filler,pointerover@parent,pointerover@host,pointerenter@host,pointerenter@parent,pointerup@parent,pointerup@host,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerout@host,pointerleave@parent,pointerleave@host"
[FAIL] mouse events with remove-slot at pointerup
  assert_equals: events received after removal expected "pointerout@filler,pointerleave@filler,pointerover@parent,pointerover@host,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerout@host,pointerleave@parent,pointerleave@host" but got "pointerout@filler,pointerleave@filler,pointerover@parent,pointerover@host,pointerenter@host,pointerenter@parent,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerout@host,pointerleave@parent,pointerleave@host"
[FAIL] mouse events with remove-filler at pointerup
  assert_equals: events received after removal expected "pointerover@slot,pointerdown@slot,pointerup@slot,pointerout@slot,pointerleave@slot,pointerleave@parent,pointerleave@host" but got "pointerover@slot,pointerenter@parent,pointerenter@slot,pointerout@slot,pointerleave@slot,pointerover@parent,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerout@host,pointerleave@parent,pointerleave@host"
[FAIL] mouse events with change-slotname at pointerup
  assert_equals: events received after removal expected "pointerout@filler,pointerleave@filler,pointerover@parent,pointerover@host,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerout@host,pointerleave@parent,pointerleave@host" but got "pointerout@filler,pointerleave@filler,pointerover@parent,pointerover@host,pointerenter@host,pointerenter@parent,pointerdown@parent,pointerdown@host,pointerup@parent,pointerup@host,pointerout@parent,pointerout@host,pointerleave@parent,pointerleave@host"
Harness: the test ran to completion.
